Customer Churn Prediction using ANN/Deployment/requirements.txt
streamlit 
pandas 
seaborn
matplotlib
plotly
Pillow
numpy
scipy
scikit-learn==1.2.1